% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{compute_ablation}
\alias{compute_ablation}
\title{Conduct an ablation analysis with a \code{caret_stack} object}
\usage{
compute_ablation(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Conduct an ablation analysis with a \code{caret_stack} object
}
\keyword{internal}
