% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{compute_feature_contributions}
\alias{compute_feature_contributions}
\title{Compute feature level contributions for a \code{caret_stack} object}
\usage{
compute_feature_contributions(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Compute feature level contributions for a \code{caret_stack} object
}
