% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{compute_metric}
\alias{compute_metric}
\title{Compute metrics for a \code{caret_stack} object}
\usage{
compute_metric(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Compute metrics for a \code{caret_stack} object
}
