% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{extract_metric.caret_list}
\alias{extract_metric.caret_list}
\title{Extract accuracy metrics from a \code{caret_list} object}
\usage{
\method{extract_metric}{caret_list}(caret_list, metric = NULL)
}
\arguments{
\item{metric}{a character string representing the metric to extract from each model.
If NULL or if the provided metric is not found, default model metrics will be extracted}

\item{x}{a \code{caret_list} object}
}
\value{
A data.table with metrics from each model.
}
\description{
Extract the cross-validated accuracy metrics from each model in a \code{caret_list}.
}
