% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{extract_metric.caret_stack}
\alias{extract_metric.caret_stack}
\title{Extract metrics from a \code{caret_stack} object}
\usage{
\method{extract_metric}{caret_stack}(caret_stack, metric = NULL)
}
\value{
A \code{data.table::data.table} of metrics of the individual and emsembled model
}
\description{
Extract metrics from a \code{caret_stack} object
}
