% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart_failure_stack}
\alias{heart_failure_stack}
\title{Pre-trained \code{caret_stack} on Heart Failure Datasets}
\format{
A \code{caret_stack} object
}
\usage{
heart_failure_stack
}
\description{
A \code{caret_stack} object pre-trained on \code{heart_failure_datasets}.
Used in examples throughout the caretMultimodal package.
}
\seealso{
\code{\link{heart_failure_datasets}}
}
\keyword{datasets}
