% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_gestational_age}
\alias{load_gestational_age}
\title{Load the example gestational age datasets}
\usage{
load_gestational_age()
}
\value{
Invisibly returns the names of the objects loaded.
}
\description{
Loads example datasets included with the package into the user's environment.
}
