% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_heart_failure}
\alias{load_heart_failure}
\title{Load the example heart failure datasets}
\usage{
load_heart_failure()
}
\value{
Invisibly returns the names of the objects loaded.
}
\description{
Loads example datasets included with the package into the user's environment.
}
\references{
Singh et al. Ensembling Electrical and Proteogenomics Biomarkers for Improved Prediction of Cardiac-Related 3-Month Hospitalizations: A Pilot Study. Can J Cardiol. 2019 Apr
}
