% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{oof_predictions}
\alias{oof_predictions}
\title{Extract out-of-fold predictions from \code{caret_list} or \code{caret_stack} objects}
\usage{
oof_predictions(object, ...)
}
\arguments{
\item{object}{A \code{caret_list} or \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Extract out-of-fold predictions from \code{caret_list} or \code{caret_stack} objects
}
\keyword{internal}
