% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{plot.caret_list}
\alias{plot.caret_list}
\title{Plot a \code{caret_list} object}
\usage{
\method{plot}{caret_list}(x, ...)
}
\arguments{
\item{x}{a caret_list object}

\item{...}{Additional arguments}
}
\value{
A \code{ggplot} object
}
\description{
This function plots the performance of each model in a caret_list object.
}
