% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{plot.caret_stack}
\alias{plot.caret_stack}
\title{Plot various metrics of a \code{caret_stack} object}
\usage{
\method{plot}{caret_stack}(x, ...)
}
\arguments{
\item{x}{The \code{caret_stack} object to plot metrics for}

\item{...}{Additional arguments}
}
\value{
A \code{ggplot} object
}
\description{
Plot various metrics of a \code{caret_stack} object
}
