% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{plot_ablation.caret_stack}
\alias{plot_ablation.caret_stack}
\title{Plot the results of an ablation analysis for a caret_stack model.}
\usage{
\method{plot_ablation}{caret_stack}(object, metric_function, metric_name, reverse = FALSE, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object.}

\item{metric_function}{A function that takes two arguments \verb{(predictions, target)}
and returns a single numeric value representing the metric to compute (e.g., RMSE, accuracy, AUC).}

\item{metric_name}{The name of the metric}

\item{reverse}{The direction to ablate in. If \code{FALSE}, the lowest contributing model is removed at each iteration.
If \code{TRUE}, the highest contributing model is removed. Default is \code{FALSE}.}

\item{...}{Not used. Included for S3 compatibility.}
}
\value{
A \code{data.table}
}
\description{
Contructs a bar plot with the output of the \code{compute_ablation} method.
}
\examples{
# Load pre-trained example caret_stack object
data(heart_failure_stack)

# Since the example stack is a binary classifier,
# this metric function needs to take in predictions (floats) and
# ground truth (binary vector), and produce a single number.
metric_fun <- function(preds, target) {
  pROC::roc(response = target, predictor = preds, quiet = TRUE)$auc
}

plot_ablation(heart_failure_stack, metric_fun, "AUC")
}
