% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{plot_ablation}
\alias{plot_ablation}
\title{Plot an ablation analysis with a \code{caret_stack} object}
\usage{
plot_ablation(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Plot an ablation analysis with a \code{caret_stack} object
}
\keyword{internal}
