% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{plot_feature_contributions.caret_stack}
\alias{plot_feature_contributions.caret_stack}
\title{Make a bar plot of feature level  for a \code{caret_stack}.}
\usage{
\method{plot_feature_contributions}{caret_stack}(object, n_features = 20, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object.}

\item{n_features}{The maximum number of features to include. Setting to a very
large value will include all features. Default is 20.}

\item{...}{Not used. Included for S3 compatibility.}
}
\value{
A \code{ggplot2} bar plot.
}
\description{
Constructs a bar plot with the output of \code{\link{compute_feature_contributions.caret_stack}}.
}
\examples{
# Load pre-trained example caret_stack object
data(heart_failure_stack)

plot_feature_contributions(heart_failure_stack)
}
