% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{plot_feature_contributions}
\alias{plot_feature_contributions}
\title{Plot feature level contributions for a \code{caret_stack} object}
\usage{
plot_feature_contributions(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Plot feature level contributions for a \code{caret_stack} object
}
