% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot metrics for a \code{caret_stack} object}
\usage{
plot_metric(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Plot metrics for a \code{caret_stack} object
}
