% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{plot_model_contributions.caret_stack}
\alias{plot_model_contributions.caret_stack}
\title{Plot the relative contributions of each of the base models in the ensemble model}
\usage{
\method{plot_model_contributions}{caret_stack}(object, descending = TRUE, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{descending}{Whether to sort in descending order. If \code{FALSE}, the output is sorted in ascending order. Default is \code{TRUE}.}

\item{...}{Additional arguments}
}
\value{
A \code{ggplot2} bar chart
}
\description{
The relative contributions are calculated using the \code{caret::varImp} function on the ensemble model.
A scaling factor is applied to make the contributions sum to 100\%.
}
