% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{plot_model_contributions}
\alias{plot_model_contributions}
\title{Plot base model contributions for a \code{caret_stack} object}
\usage{
plot_model_contributions(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Plot base model contributions for a \code{caret_stack} object
}
\keyword{internal}
