% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_registrations.R
\name{plot_roc}
\alias{plot_roc}
\title{Make an ROC plot for a \code{caret_stack} object}
\usage{
plot_roc(object, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{...}{Additional arguments passed to class-specific methods}
}
\description{
Make an ROC plot for a \code{caret_stack} object
}
\keyword{internal}
