% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{predict.caret_list}
\alias{predict.caret_list}
\title{Create a matrix of predictions for each model in a caret_list}
\usage{
\method{predict}{caret_list}(object, data_list, drop_redundant_class = TRUE, ...)
}
\arguments{
\item{object}{A \code{caret_list} object}

\item{data_list}{A list of datasets to predict on, with each dataset matching the corresponding model in \code{caret_list}.}

\item{drop_redundant_class}{A boolean controlling whether to exclude the first class level from prediction output. Default is \code{TRUE}.}

\item{...}{Additional arguments to pass to \code{caret::predict}}
}
\value{
A \code{data.table::data.table} of predictions
}
\description{
This always return probabilities for classification models, with the option to drop one predicted class.
}
