% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{predict.caret_stack}
\alias{predict.caret_stack}
\title{Create a matrix of predictions for a \code{caret_stack} object.}
\usage{
\method{predict}{caret_stack}(object, data_list, drop_redundant_class = TRUE, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{data_list}{A list of datasets to predict on, with each dataset matching the corresponding model in \code{caret_list}.}

\item{drop_redundant_class}{A boolean controlling whether to exclude the first class from prediction output. Default is \code{TRUE}.}

\item{...}{Additional arguments to pass to \code{caret::predict}.}
}
\value{
A \code{data.table::data.table} of predictions for base and ensemble models.
}
\description{
Create a matrix of predictions for a \code{caret_stack} object.
}
\examples{
# Load example data and pre-trained caret_stack object
data(heart_failure_datasets)
data(heart_failure_stack)

# List of datasets to predict on
data_list <- heart_failure_datasets[c("cells", "holter", "mrna", "proteins")]

predict(heart_failure_stack, data_list)
}
