% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mae.R
\name{prepare_mae}
\alias{prepare_mae}
\title{Prepare a MultiAssayExperiment for use with caretMultimodal}
\usage{
prepare_mae(mae, transpose = FALSE, ...)
}
\arguments{
\item{mae}{The MultiAssayExperiment object.}

\item{transpose}{Whether to transpose the individual matrices. Samples must correspond to rows for caret_list.
Default is FALSE.}

\item{...}{Not used. Included for S3 compatibility.}
}
\value{
A named list of matrices.
}
\description{
Converts a MultiAssayExperiment object from the MultiAssayExperiment package to a
simple list of datasets to pass into \code{caret_list}.
}
