% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{print.caret_stack}
\alias{print.caret_stack}
\title{Print details of a \code{caret_stack} object.}
\usage{
\method{print}{caret_stack}(x, ...)
}
\arguments{
\item{x}{A \code{caret_stack} object}

\item{...}{Additional arguments}
}
\description{
Print details of a \code{caret_stack} object.
}
