% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{print.summary.caret_list}
\alias{print.summary.caret_list}
\title{Print a summary of a caret_list}
\usage{
\method{print}{summary.caret_list}(x, ...)
}
\arguments{
\item{x}{A \code{summary.caret_list} object}

\item{...}{Additional arguments}
}
\description{
Print a summary of a caret_list
}
