% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{print.summary.caret_stack}
\alias{print.summary.caret_stack}
\title{Print a summary of a \code{caret_stack} object}
\usage{
\method{print}{summary.caret_stack}(summary)
}
\description{
Print a summary of a \code{caret_stack} object
}
