% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{summary.caret_stack}
\alias{summary.caret_stack}
\title{Get a summary of a \code{caret_stack} object}
\usage{
\method{summary}{caret_stack}(caret_stack)
}
\description{
Get a summary of a \code{caret_stack} object
}
