% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{caret_list}
\alias{caret_list}
\title{Create a \code{caret_list} object}
\usage{
caret_list(
  target,
  data_list,
  method,
  trControl = NULL,
  metric = NULL,
  continue_on_fail = FALSE,
  trim = TRUE,
  aggregate_resamples = TRUE,
  ...
)
}
\arguments{
\item{target}{Target parameter vector, either numeric for regression or a factor/character for classification.}

\item{data_list}{A list of data sets to train models on}

\item{method}{The method to train the models with. Can be a custom method or one found in \code{caret::modelLookup()}.}

\item{trControl}{Control for use with the \code{caret::train} function. A default control will be constructed depending on the target type.}

\item{metric}{Metric for use with \code{caret::train} function. A default metric will be constructed depending on the target type.}

\item{continue_on_fail}{Logical, whether to skip over a data set if the model fails to train. Default is \code{FALSE}.}

\item{trim}{Logical, whether the train models be trimmed to save memory and speed up stacking.}

\item{aggregate_resamples}{Logical, whether to aggregate stacked predictions.}

\item{...}{Any additional arguments to pass to the \code{caret::train} function}
}
\value{
A \code{caret_list} object, which is a list of \code{caret::train} model corresponding to \code{data_list}.
}
\description{
Build a list of \code{caret::train} objects, where each model corresponds to a data set in \code{data_list}.
The resulting list is structured for use in ensembling workflows. Users can specify the training method, control parameters,
and metrics, and the function supports error handling and model trimming for efficiency.
}
