% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{caret_list}
\alias{caret_list}
\title{Create a \code{caret_list} object}
\usage{
caret_list(
  target,
  data_list,
  method,
  identifier_column_name = NULL,
  trControl = NULL,
  metric = NULL,
  trim = TRUE,
  ...
)
}
\arguments{
\item{target}{Target parameter vector, either numeric for regression or a factor/character for classification.}

\item{data_list}{A list of data sets to train models on}

\item{method}{The method to train the models with. Can be a custom method or one found in \code{caret::modelLookup()}.}

\item{identifier_column_name}{The name of a column that connects the rows in the dataset (ex. a participant ID).
If provided, this column must be present in all datasets within the \code{data_list} for proper matching. Use this if the datasets have different numbers of rows.}

\item{trControl}{Control for use with the \code{caret::train} function.
If \code{NULL}, a default control will be constructed depending on the target type.}

\item{metric}{Metric for use with \code{caret::train} function.
If \code{NULL}, a default metric will be constructed depending on the target type.}

\item{trim}{Logical, whether the train models be trimmed to save memory. Default is \code{TRUE}}

\item{...}{Any additional arguments to pass to the \code{caret::train} function}

\item{aggregate_resamples}{Whether to aggregate out-of-fold predictions across multiple resamples. Default is \code{TRUE}.}
}
\value{
A \code{caret_list} object, which is a list of \code{caret::train} model corresponding to \code{data_list}.
}
\description{
Build a list of \code{caret::train} objects, where each model corresponds to a data set in \code{data_list}.
The resulting list is structured for use in ensembling workflows. Users can specify the training method, control parameters,
and metrics, and the function supports error handling and model trimming for efficiency.
}
