% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{caret_stack}
\alias{caret_stack}
\title{Ensemble the models of a \code{caret_list} object}
\usage{
caret_stack(
  caret_list,
  method,
  target = NULL,
  new_data_list = NULL,
  metric = NULL,
  trControl = NULL,
  excluded_class_id = 1L,
  aggregate_resamples = TRUE
)
}
\arguments{
\item{caret_list}{a \code{caret_list} object}

\item{method}{The method to train the ensemble model. Can be a custom method or one found in \code{caret::modelLookup()}.}

\item{target}{Target parameter vector. Defaults to \code{NULL}, in which case the training data from the first model is \code{caret_list} will be used.}

\item{new_data_list}{A list of datasets to predict on, with each dataset matching the corresponding model in \code{caret_list}. Defaults to \code{NULL}, in which case the training data for each model is used.}

\item{metric}{Metric for use with \code{caret::train} function. A default metric will be constructed depending on the target type.}

\item{trControl}{Control for use with the \code{caret::train} function. A default control will be constructed depending on the target type.}

\item{excluded_class_id}{An integer indicating the class to exclude from predictions. If 0L, no class is excluded. Default is 1L.}

\item{aggregate_resamples}{Boolean, whether to aggregate resamples by keys. Default is \code{TRUE}.}
}
\value{
A \code{caret_stack} object.
}
\description{
Stack several \code{caret::train} models from a \code{caret_list} object using a \code{caret::train} model.
}
