% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{compute_feature_contributions.caret_stack}
\alias{compute_feature_contributions.caret_stack}
\title{Compute the feature level contributions for a caret_stack model.}
\usage{
\method{compute_feature_contributions}{caret_stack}(object, n_features = 20, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object.}

\item{n_features}{The maximum number of features to include. Setting to a very
large value will include all features. Default is 20.}

\item{...}{Not used. Included for S3 compatibility.}
}
\value{
A \code{data.table}
}
\description{
Feature-level contributions for the ensemble model are
computed using a two-stage application of caret::varImp. First,
varImp is applied to the ensemble model, where the base-model predictions
are treated as features, yielding dataset-level weights. Next, varImp is
applied to each base model to obtain feature-level importance scores within
each dataset. The final contribution of an individual feature to the ensemble
is calculated as the product of its dataset-level weight and its feature-level
importance within the corresponding base model.
}
