% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{compute_metric.caret_stack}
\alias{compute_metric.caret_stack}
\title{Compute metrics with a provided metric function}
\usage{
\method{compute_metric}{caret_stack}(object, metric_function, metric_name, descending = TRUE, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{metric_function}{A function that takes two arguments \verb{(predictions, target)}
and returns a single numeric value representing the metric to compute (e.g., RMSE, accuracy, AUC).}

\item{metric_name}{The name of the metric}

\item{descending}{Whether to sort in descending order. If \code{FALSE}, the output is sorted in ascending order. Default is \code{TRUE}.}

\item{...}{Additional arguments}
}
\value{
A \code{data.table} of metrics
}
\description{
The metric_function is applied to the out-of-fold predictions for the caret_stack.
}
