% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{extract_metric.caret_stack}
\alias{extract_metric.caret_stack}
\title{Extract metrics from a \code{caret_stack} object}
\usage{
\method{extract_metric}{caret_stack}(x, ...)
}
\arguments{
\item{x}{A \code{caret_stack} object}

\item{...}{Additional arguments}
}
\value{
A \code{data.table::data.table} of metrics
}
\description{
Produce metrics for how the caret_list models and ensemble model perform on the training data for the caret stack
}
