% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart_failure_datasets}
\alias{heart_failure_datasets}
\title{Heart Failure Datasets}
\format{
A named list with 5 elements:
\describe{
\item{demo}{A \code{data.frame} of demographic features}
\item{cells}{A \code{data.frame} of cell features}
\item{holter}{A \code{data.frame} of Holter monitor features}
\item{mrna}{A \code{data.frame} of mRNA features}
\item{proteins}{A \code{data.frame} of protein features}
}
}
\source{
Singh et al. Ensembling Electrical and Proteogenomics Biomarkers for Improved
Prediction of Cardiac-Related 3-Month Hospitalizations: A Pilot Study. Can J Cardiol. 2019 Apr
}
\usage{
heart_failure_datasets
}
\description{
This is a sample description I will fix later
}
\keyword{datasets}
