% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{oof_predictions.caret_list}
\alias{oof_predictions.caret_list}
\title{Out-of-fold predictions from a caret_list}
\usage{
\method{oof_predictions}{caret_list}(
  object,
  drop_redundant_class = TRUE,
  aggregate_resamples = TRUE,
  intersection_only = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{caret_list} object}

\item{drop_redundant_class}{A boolean controlling whether to exclude the first class level from prediction output. Default is \code{TRUE}.}

\item{aggregate_resamples}{Logical, whether to aggregate resamples across folds.}

\item{intersection_only}{Logical, whether to trim down the out of fold predictions to only the intersection of
samples that have data in all datasets.}

\item{...}{Not used. Included for S3 compatibility.}
}
\value{
A \code{data.table::data.table} of OOF predictions
}
\description{
Retrieve the out-of-fold predictions corresponding to the best
hyperparameter setting of a trained caret model. These predictions come from
the resampling process (not the final refit) and can optionally be aggregated
across resamples to produce a single prediction per training instance.
TODO touch up description
}
