% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{oof_predictions.caret_list}
\alias{oof_predictions.caret_list}
\title{Out-of-fold predictions from a caret_list}
\usage{
\method{oof_predictions}{caret_list}(
  object,
  excluded_class_id = 1L,
  aggregate_resamples = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{caret_list} object}

\item{excluded_class_id}{An integer indicating the class index to exclude from prediction output.
If \code{NONE}, no class is excluded. Default is 1L.}

\item{aggregate_resamples}{Logical, whether to aggregate resamples across folds.}

\item{...}{Additional arguments}
}
\value{
A \code{data.table::data.table} of OOF predictions
}
\description{
Retrieve the out-of-fold predictions corresponding to the best
hyperparameter setting of a trained caret model. These predictions come from
the resampling process (not the final refit) and can optionally be aggregated
across resamples to produce a single prediction per training instance.
}
