% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{plot.caret_stack}
\alias{plot.caret_stack}
\title{Plot metrics of a \code{caret_stack} object}
\usage{
\method{plot}{caret_stack}(caret_stack, metric = NULL)
}
\value{
A \code{ggplot2} object
}
\description{
Plot metrics of a \code{caret_stack} object
}
