% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{plot_metric.caret_stack}
\alias{plot_metric.caret_stack}
\title{Plot metrics computed with a provided metric function}
\usage{
\method{plot_metric}{caret_stack}(object, metric_function, metric_name, descending = TRUE, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{metric_function}{A function that takes two arguments \verb{(predictions, target)}
and returns a single numeric value representing the metric to compute (e.g., RMSE, accuracy, AUC).}

\item{metric_name}{The name of the metric}

\item{descending}{Whether to sort in descending order. If \code{FALSE}, the output is sorted in ascending order. Default is \code{TRUE}.}

\item{...}{Additional arguments}
}
\value{
A \code{ggplot2} bar chart
}
\description{
This function constructs a bar plot with the output of the compute metric method.
The bars are ordered by increasing value.
}
