% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{plot_roc.caret_stack}
\alias{plot_roc.caret_stack}
\title{Plot ROC curves for individual and ensemble models in a caret_stack}
\usage{
\method{plot_roc}{caret_stack}(object, include_auc = TRUE, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{include_auc}{Whether to include AUC values in the legend. Default is \code{True}.}

\item{...}{Additional arguments}
}
\value{
A \code{ggplot2} object
}
\description{
This function calculates ROC curves for all base models and the ensemble model
using the out-of-fold predictions from a \code{caret_stack} object.
The \code{pROC} package is used to compute the ROC curves. ROC curves can only be constructed for binary calssifiers.
}
