% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{predict.caret_list}
\alias{predict.caret_list}
\title{Create a matrix of predictions for each model in a caret_list}
\usage{
\method{predict}{caret_list}(object, data_list, excluded_class_id = 1L, ...)
}
\arguments{
\item{object}{A \code{caret_list} object}

\item{data_list}{A list of datasets to predict on, with each dataset matching the corresponding model in \code{caret_list}.}

\item{excluded_class_id}{An integer indicating the class index to exclude from prediction output.
If \code{NULL}, no class is excluded. Default is 1L.}

\item{...}{Additional arguments to pass to \code{caret::predict}}
}
\value{
A \code{data.table::data.table} of predictions
}
\description{
This always return probabilities for classification models, with the option to drop one predicted class.
}
