% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{predict.caret_list}
\alias{predict.caret_list}
\title{Create a matrix of predictions for each model in a caret_list}
\usage{
\method{predict}{caret_list}(
  caret_list,
  new_data_list = NULL,
  verbose = FALSE,
  excluded_class_id = 1L,
  aggregate_resamples = TRUE
)
}
\arguments{
\item{caret_list}{A \code{caret_list} object}

\item{new_data_list}{A list of datasets to predict on, with each dataset matching the corresponding model in \code{caret_list}. If \code{NULL}, predictions from the training data are returned.}

\item{verbose}{A boolean that controls the verbosity of error messages}

\item{excluded_class_id}{An integer indicating the class to exclude from predictions. If 0L, no class is excluded. Default is 1L.}

\item{aggregate_resamples}{A boolean that controls whether to aggregate resamples by keys. Default is \code{TRUE}.}
}
\value{
A \code{data.table::data.table} of predictions
}
\description{
Create a matrix of predictions for each model in a caret_list
}
