% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{predict.caret_stack}
\alias{predict.caret_stack}
\title{Create a matrix of predictions for a \code{caret_stack} object.}
\usage{
\method{predict}{caret_stack}(object, data_list, excluded_class_id = 1L, ...)
}
\arguments{
\item{object}{A \code{caret_stack} object}

\item{data_list}{A list of datasets to predict on, with each dataset matching the corresponding model in \code{caret_list}.}

\item{excluded_class_id}{An integer indicating the class index to exclude from prediction output.
If \code{NULL}, no class is excluded. Default is 1L.}

\item{...}{Additional arguments}
}
\value{
A \code{data.table::data.table} of predictions for base and ensemble models.
}
\description{
Create a matrix of predictions for a \code{caret_stack} object.
}
