% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{predict.caret_stack}
\alias{predict.caret_stack}
\title{Create a matrix of predictions for a \code{caret_stack} object.}
\usage{
\method{predict}{caret_stack}(
  caret_stack,
  new_data_list = NULL,
  verbose = FALSE,
  excluded_class_id = 1L,
  aggregate_resamples = TRUE
)
}
\arguments{
\item{caret_stack}{A \code{caret_stack} object}

\item{new_data_list}{New data to predict on. Default is \code{NULL}, in which case the training data is predicted on.}

\item{verbose}{Boolean that controls the verbosity of error messages}

\item{excluded_class_id}{An integer indicating the class to exclude from predictions. If 0L, no class is excluded. Default is 1L.}

\item{aggregate_resamples}{Boolean, controls whether to aggregate resamples by keys. Default is \code{TRUE}.}
}
\value{
A \code{data.table::data.table} of predictions
}
\description{
Create a matrix of predictions for a \code{caret_stack} object.
}
