% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_stack.R
\name{print.summary.caret_stack}
\alias{print.summary.caret_stack}
\title{Print a summary of a \code{caret_stack} object}
\usage{
\method{print}{summary.caret_stack}(x, ...)
}
\arguments{
\item{x}{A \code{summary.caret_stack} object}

\item{...}{Additional arguments}
}
\description{
Print a summary of a \code{caret_stack} object
}
