% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_list.R
\name{summary.caret_list}
\alias{summary.caret_list}
\title{Summarize a caret_list}
\usage{
\method{summary}{caret_list}(object, ...)
}
\arguments{
\item{object}{a \code{caret_list} object}

\item{...}{Additional arguments}
}
\value{
A \code{summary.caret_list} object
}
\description{
This function summarizes the performance of each model in a caret_list object.
}
